Module Meja_lib.Envi

module IdTbl = Ident.Table

type error =

| No_open_scopes

| Wrong_scope_kind of string

| Multiple_definition of string * string

| Unbound_type_var of Type0.type_expr

| Unbound_type of Ast_types.Longident.t

| Unbound_module of Ast_types.Longident.t

| Unbound_value of Ast_types.Longident.t

| Unbound of string * Ast_types.Longident.t

| Unbound_path of string * Path.t

| Wrong_number_args of Path.t * int * int

| Expected_type_var of Type0.type_expr

| Lident_unhandled of string * Ast_types.Longident.t

| Constraints_not_satisfied of Type0.type_expr * Type0.type_decl

| No_unifiable_implicit

| Recursive_load of string

| Predeclared_types of Ident.t list

| Functor_in_module_sig

| Not_a_functor

| Duplicate_name of string * bool * string

| Shadowed_checked of string * Ast_types.Longident.t

exception Error of Compiler_internals.Location.t * error

module TypeEnvi : sig ... end

type 'a or_deferred =

| Immediate of 'a

| Deferred of string

| In_flight of string

type 'a resolve_env = {

mutable type_env : TypeEnvi.t;

mutable external_modules : 'a or_deferred IdTbl.t;

}

module Shadows : sig ... end

val reject_shadowed : loc:Compiler_internals.Location.t -> string -> Ast_types.Longident.t -> Ast_types.mode -> Ast_types.mode -> bool

Rejecting mode checker. If the mode is checked but a prover-mode name is found first, raise an error complaining about shadowing. When in prover mode, this accepts names from either mode.

module Scope : sig ... end

val empty_resolve_env : Scope.t resolve_env

type t = {

scope_stack : Scope.t list;

depth : int;

resolve_env : Scope.t resolve_env;

}

val empty : Scope.t resolve_env -> t

val current_scope : t -> Scope.t

val push_scope : Scope.t -> t -> t

val current_mode : t -> Ast_types.mode

val current_shadow : t -> Shadows.t Core_kernel.ref

val mode_or_default : Ast_types.mode option -> t -> Ast_types.mode

val make_functor : mode:Ast_types.mode -> (Scope.t Scope.or_path -> Scope.t) -> Scope.t

val open_expr_scope : ?mode:Ast_types.mode -> t -> t

val open_module : ?mode:Ast_types.mode -> t -> t

val open_namespace_scope : ?mode:Ast_types.mode -> Path.t -> Scope.t -> t -> t

val open_instance_scope : ?mode:Ast_types.mode -> Path.t -> Scope.t -> t -> t

val open_mode_module_scope : Ast_types.mode -> t -> t

val pop_scope : t -> Scope.t * t

val pop_expr_scope : t -> Scope.t * t

val pop_module : loc:Compiler_internals.Location.t -> t -> Scope.t * t

val close_expr_scope : t -> t

val map_current_scope : f:(Scope.t -> Scope.t) -> t -> t

val get_var_names : t -> (Type0.type_expr, Core_kernel.String.Map.Key.t, Typeset.Map.Key.comparator_witness) Core_kernel.Map.t * (Core_kernel.String.Map.Key.t, Core_kernel.String.Set.Elt.comparator_witness) Core_kernel.Set.t

val set_var_names : t -> Type0_map.mapper

val add_type_variable : Core_kernel.String.Map.Key.t -> Type0.type_expr -> t -> t

val find_type_variable : mode:Ast_types.mode -> Core_kernel.String.Map.Key.t -> t -> Type0.type_expr option

val add_module : loc:Compiler_internals.Location.t -> Ident.t -> Scope.t -> t -> t

val add_deferred_module : loc:Compiler_internals.Location.t -> Ident.t -> string -> t -> t

val register_external_module : Ident.t -> Scope.t or_deferred -> t -> unit

val fake_alias_external_module_exn : fake_name:Ident.t -> string -> t -> unit

val find_module : loc:Compiler_internals.Location.t -> Ast_types.lid -> t -> mode:Ast_types.mode -> Path.t * Scope.t

val find_module_deferred : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> Ast_types.lid -> t -> (Path.t * Scope.t Scope.or_path) option

val add_implicit_instance : Ident.t -> Type0.type_expr -> t -> t

val find_of_lident : mode:Ast_types.mode -> kind:string -> get_name:(modes:(Ast_types.mode -> bool) -> string -> Scope.t -> (Ident.t * 'a) Base__Option.t) -> Ast_types.lid -> t -> (Path.t * 'a) option

val get_of_path : loc:Compiler_internals.Location.t -> kind:string -> get_name:(Ident.t -> Scope.t -> 'a option) -> find_name:(modes:(Ast_types.mode -> bool) -> string -> Scope.t -> (Ident.t * 'a) option) -> Path.t -> t -> Ident.t * 'a

val join_expr_scope : t -> Scope.t -> t

val has_type_declaration : mode:Ast_types.mode -> Ast_types.lid -> t -> bool

val raw_find_type_declaration : mode:Ast_types.mode -> Ast_types.lid -> t -> Path.t * Type0.type_decl

val raw_get_type_declaration : loc:Compiler_internals.Location.t -> Path.t -> t -> Ident.t * Type0.type_decl

val add_module_type : loc:Compiler_internals.Location.t -> Ident.t -> Scope.t Scope.or_path -> t -> t

val find_module_type : Ast_types.lid -> mode:Ast_types.mode -> t -> (Path.t * Scope.t Scope.or_path) option

module Type : sig ... end

module TypeDecl : sig ... end

val add_name : Meja_lib__Ident.t -> Type0.type_expr -> t -> t

val find_name : mode:Ast_types.mode -> Ast_types.lid -> t -> Path.t * Type0.type_expr

val get_of_constructor : loc:Compiler_internals.Location.t -> Path.t -> t -> Ident.t * (Type0.type_decl * int)

val find_conversion : unifies:(Type.env -> Type0.type_expr -> Type0.type_expr -> bool) -> Type0.type_expr -> Type.env -> (Path.t * (Ast_types.arg_label * Type0.type_expr) Core_kernel.List.t) option

val wrap_prover_implicits : t -> unit

Wrap any prover implicit arguments with Tprover, so that they may be surfaced as implicit arguments within checked types.

val pp_typ : Stdlib.Format.formatter -> Type0.type_expr -> unit

val pp_decl_typ : Stdlib.Format.formatter -> Type0.type_decl -> unit

val report_error : Stdlib.Format.formatter -> error -> unit