Module Envi.Scope

type 'a or_path =

| Immediate of 'a

| Deferred of string

type 't kind =

| Module of Shadows.t Core_kernel.ref

| Expr

| Open of Path.t

| Open_instance of Path.t

| Continue of Shadows.t Core_kernel.ref

| Functor of 't or_path -> 't

type t = {

kind : t kind;

names : Type0.type_expr IdTbl.t;

type_variables : Type0.type_expr Core_kernel.String.Map.t;

type_decls : Type0.type_decl IdTbl.t;

fields : (Type0.type_decl * int) IdTbl.t;

ctors : (Type0.type_decl * int) IdTbl.t;

modules : t or_path IdTbl.t;

module_types : t or_path IdTbl.t;

instances : (Path.t * Path.t * Type0.type_expr) list;

mode : Ast_types.mode;

}

val load_module : (loc:Compiler_internals.Location.t -> name:string -> t resolve_env -> string -> t) Core_kernel.ref

val empty : mode:Ast_types.mode -> t kind -> t

val get_shadow : t -> Shadows.t Core_kernel.ref

val add_name : Meja_lib__Ident.t -> Type0.type_expr -> t -> t

val get_name : Meja_lib__Ident.t -> t -> Type0.type_expr option

val find_name : modes:(Ast_types.mode -> bool) -> string -> t -> (Meja_lib__Ident.t * Type0.type_expr) option

val add_type_variable : Core_kernel.String.Map.Key.t -> Type0.type_expr -> t -> t

val find_type_variable : mode:Ast_types.mode -> Core_kernel.String.Map.Key.t -> t -> Type0.type_expr Core_kernel.Option.t

val add_field : Type0.type_decl -> int -> t -> Type0.field_decl -> t

val get_field : Meja_lib__Ident.t -> t -> (Type0.type_decl * int) option

val find_field : mode:Ast_types.mode -> modes:'a -> string -> t -> (Meja_lib__Ident.t * (Type0.type_decl * int)) option

val add_ctor : Type0.type_decl -> int -> t -> Type0.ctor_decl -> t

val get_ctor : Meja_lib__Ident.t -> t -> (Type0.type_decl * int) option

val find_ctor : mode:Ast_types.mode -> modes:'a -> string -> t -> (Meja_lib__Ident.t * (Type0.type_decl * int)) option

val add_type_declaration : loc:Compiler_internals.Location.t -> ?may_shadow:bool -> Ident.t -> Type0.type_decl -> t -> t

val get_type_declaration : Meja_lib__Ident.t -> t -> Type0.type_decl option

val find_type_declaration : mode:Ast_types.mode -> modes:'a -> string -> t -> (Meja_lib__Ident.t * Type0.type_decl) option

val register_type_declaration : loc:Compiler_internals.Location.t -> ?may_shadow:bool -> Ident.t -> Type0.type_decl -> t -> t

val fold_over : init:'a -> names:(key:string -> data:[ `Both of Type0.type_expr * Type0.type_expr | `Left of Type0.type_expr | `Right of Type0.type_expr ] -> 'b -> 'b) -> type_variables:(key:Core_kernel.String.Map.Key.t -> data:[ `Both of Type0.type_expr * Type0.type_expr | `Left of Type0.type_expr | `Right of Type0.type_expr ] -> 'a -> 'a) -> type_decls:(key:string -> data:[ `Both of Type0.type_decl * Type0.type_decl | `Left of Type0.type_decl | `Right of Type0.type_decl ] -> 'a -> 'a) -> fields:(key:string -> data:[ `Both of (Type0.type_decl * int) * (Type0.type_decl * int) | `Left of Type0.type_decl * int | `Right of Type0.type_decl * int ] -> 'a -> 'a) -> ctors:(key:string -> data:[ `Both of (Type0.type_decl * int) * (Type0.type_decl * int) | `Left of Type0.type_decl * int | `Right of Type0.type_decl * int ] -> 'a -> 'a) -> modules:(key:string -> data:[ `Both of t or_path * t or_path | `Left of t or_path | `Right of t or_path ] -> 'a -> 'a) -> module_types:(key:string -> data:[ `Both of t or_path * t or_path | `Left of t or_path | `Right of t or_path ] -> 'a -> 'a) -> instances:('a -> (Path.t * Path.t * Type0.type_expr) list -> (Path.t * Path.t * Type0.type_expr) list -> 'b) -> t -> t -> 'b

val join : loc:Compiler_internals.Location.t -> t -> t -> t

val add_module : loc:Compiler_internals.Location.t -> Ident.t -> t or_path -> t -> t

val add_module_type : loc:Compiler_internals.Location.t -> Ident.t -> t or_path -> t -> t

val get_module_type : modes:(Ast_types.mode -> bool) -> string -> t -> (Meja_lib__Ident.t * t or_path) option

val outer_mod_name : loc:Compiler_internals.Location.t -> Ast_types.Longident.t -> string * Ast_types.Longident.t option

val subst : Subst.t -> t -> t

val build_subst : type_subst:(Meja_lib__Ident.t -> Path.Map.Key.t * Path.t) -> module_subst:(Meja_lib__Ident.t -> Path.Map.Key.t * Path.t) -> expr_subst:(local_path:Path.t -> Path.t -> Path.Map.Key.t * Path.t) -> Subst.t -> t -> Subst.t

val register_external_module : Ident.t -> t or_deferred -> t resolve_env -> unit

val load_external_module : loc:Compiler_internals.Location.t -> string -> Ident.t -> t resolve_env -> t

val get_global_module : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> t resolve_env -> string -> (Meja_lib__Ident.t * t) option

val find_module_ : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> scopes:t Core_kernel.List.t -> t resolve_env -> Ast_types.Longident.t -> t -> (Path.t * t) Base__Option.t

val apply_functor : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> scopes:t Core_kernel.List.t -> t resolve_env -> Path.t -> Ast_types.Longident.t -> t -> Path.t * t

val get_module : modes:(Ast_types.mode -> bool) -> loc:Compiler_internals.Location.t -> scopes:t Core_kernel.List.t -> t resolve_env -> string -> t -> (Ident.t * t) Base__Option.t

val find_global_module : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> scopes:t Core_kernel.List.t -> t resolve_env -> Ast_types.Longident.t -> (Path.t * t) Core_kernel.Option.t

val find_module : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> Ast_types.Longident.t -> t resolve_env -> t Core_kernel.List.t -> Path.t * t

val find_module_deferred : mode:Ast_types.mode -> loc:Compiler_internals.Location.t -> scopes:t Core_kernel.List.t -> t resolve_env -> Ast_types.Longident.t -> t -> (Path.t * t or_path) Base__Option.t

val join_expr_scope : t -> t -> t

val get_module_by_ident : loc:Compiler_internals.Location.t -> t resolve_env -> Ident.t -> t -> t Core_kernel.Option.t

Get the module referred to by the given Ident.t.

This skips over any modules which are not immediate. As such, this should only be used to retrieve sub-values of the module, which must have forced the loading of the module at their creation time.

val get_module_no_load : loc:Compiler_internals.Location.t -> t resolve_env -> modes:(Ast_types.mode -> bool) -> string -> t -> (Meja_lib__Ident.t * t) Core_kernel.Option.t

Get the module referred to by the given name for the given mode.

This skips over any modules which are not immediate. As such, this should only be used to retrieve sub-values of the module, which must have forced the loading of the module at their creation time.

val global : t resolve_env -> t