Module Envi.Shadows

type t = {

types : Ident.t Core_kernel.String.Map.t;

modules : Core_kernel.String.Set.t;

module_types : Core_kernel.String.Set.t;

}

A register of names already present in the current module. Used to reject duplicate names that would shadow existing path-dependent items.

val empty : t