Module Envi.Type

type env = t

val mkvar : mode:Ast_types.mode -> string option -> t -> Type0.type_expr

module Mk : sig ... end

val map_env : f:(TypeEnvi.t -> TypeEnvi.t) -> t -> unit

val refresh_vars : Type0.type_expr Core_kernel.List.t -> t -> unit

Replace the representatives of each of list of variables with a fresh variable.

The old values can be restored by taking a snapshot before calling this and backtracking to it once the new values have been used.

copy must be called on each of the values, to ensure that a tri-stitching Tvar -> _ <-> _ (for non-Tvar stitched types) is properly instantiated.

val copy : Type0.type_expr -> t -> Type0.type_expr

val instantiate : Type0.type_expr Core_kernel.List.t -> Type0.type_expr Core_kernel.List.t -> Type0.type_expr -> t -> Type0.type_expr

instantiate params typs typ env creates a new type by replacing the each of the type variables in params with the corresponding instance from typs in typ.

module T : sig ... end

val update_depths : t -> Type0.type_expr -> unit

val or_compare : Core_kernel.Int.t -> f:(unit -> Core_kernel.Int.t) -> Core_kernel.Int.t

val compare : loc:Compiler_internals.Location.t -> t -> Type0.type_expr -> Type0.type_expr -> int

val compare_all : loc:Compiler_internals.Location.t -> t -> Type0.type_expr list -> Type0.type_expr list -> Core_kernel.Int.t

val compare_row : loc:Compiler_internals.Location.t -> t -> Type0.row -> Type0.row -> Core_kernel.Int.t

val weak_variables : int -> (Type0.type_expr, 'a) Core_kernel.Set.t -> Type0.type_expr -> (Type0.type_expr, 'a) Core_kernel.Set.t

val new_implicit_var : ?loc:Compiler_internals.Location.t -> Type0.type_expr -> t -> Typedast.expression

val implicit_instance : unifies:(env -> Type0.type_expr -> Type0.type_expr -> bool) -> Type0.type_expr -> (Type0.type_expr, 'a) Core_kernel.Set.t -> env -> (Path.t * Type0.type_expr) option

val generate_implicits : Typedast.expression -> t -> Typedast.expression

val instantiate_implicits : loc:'a -> unifies:(env -> Type0.type_expr -> Type0.type_expr -> bool) -> Typedast.expression Core_kernel.List.t -> env -> Typedast.expression Core_kernel.List.t

val flattened_implicit_vars : loc:Compiler_internals.Location.t -> toplevel:bool -> unifies:(env -> Type0.type_expr -> Type0.type_expr -> bool) -> (Type0.type_expr, Typeset.Elt.comparator_witness) Core_kernel.Set.t -> env -> Typedast.expression Base__List.t