Module Meja_lib.Ident

type t

Unique identifiers.

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val equal : t -> t -> bool

val compare : t -> t -> int

type ident = t

val sexp_of_ident : ident -> Ppx_sexp_conv_lib.Sexp.t

val ident_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> ident

val create : mode:Ast_types.mode -> ?ocaml:bool -> string -> t

Create a new unique name.

If the ocaml argument is true, the identifier is associated with a modifiable 'OCaml name', accessible via ocaml_name_ref.

val create_global : string -> t

Create the name of a global (ie. external) module.

val create_row : string -> t

Create an identifier for a row constructor.

This kind of identifier will alias with any other row identifier created with the same name.

val is_global : t -> bool

val is_row : t -> bool

val name : t -> string

Retrieve the name passed to create.

val ocaml_name : t -> string

Return the OCaml name for this identifier, which may or may not differ from the name given by name.

val ocaml_name_ref : t -> string Stdlib.ref option

Returns a handle to modify the OCaml name, or none if there is no associated OCaml name.

val mode : t -> Ast_types.mode

Retrieve the mode passed to create.

val pprint : Stdlib.Format.formatter -> t -> unit

Pretty print. Identifiers that do not begin with a letter or underscore will be surrounded by parentheses.

val debug_print : Stdlib.Format.formatter -> t -> unit

Debug print. Prints the identifier and its internal ID.

val fresh : Ast_types.mode -> t

HACK: Call a variable "x___" ^ id where id is the identifier's internal id.

module Table : sig ... end

module Set : Core_kernel.Set.S with type Elt.t = t

module Map : Core_kernel.Map.S with type Key.t = t