Module Meja_lib.Parsetypes

type type_expr = {

type_desc : type_desc;

type_loc : Location.t;

}

and type_desc =

| Ptyp_var of Ast_types.str option

| Ptyp_tuple of type_expr list

| Ptyp_arrow of type_expr * type_expr * Ast_types.explicitness * Asttypes.arg_label

| Ptyp_ctor of variant

| Ptyp_poly of type_expr list * type_expr

| Ptyp_prover of type_expr

| Ptyp_conv of type_expr * type_expr

| Ptyp_opaque of type_expr

| Ptyp_alias of type_expr * Ast_types.str

| Ptyp_row of row_tag list * Ast_types.closed_flag * Ast_types.str list option

| Ptyp_row_subtract of type_expr * Ast_types.str list

and variant = {

var_ident : Ast_types.lid;

var_params : type_expr list;

}

and row_tag = {

rtag_ident : Ast_types.str;

rtag_arg : type_expr list;

rtag_loc : Location.t;

}

type field_decl = {

fld_ident : Ast_types.str;

fld_type : type_expr;

fld_loc : Location.t;

}

type ctor_args =

| Ctor_tuple of type_expr list

| Ctor_record of field_decl list

type ctor_decl = {

ctor_ident : Ast_types.str;

ctor_args : ctor_args;

ctor_ret : type_expr option;

ctor_loc : Location.t;

}

type type_decl = {

tdec_ident : Ast_types.str;

tdec_params : type_expr list;

tdec_desc : type_decl_desc;

tdec_loc : Location.t;

}

and type_decl_desc =

| Pdec_abstract

| Pdec_alias of type_expr

| Pdec_record of field_decl list

| Pdec_variant of ctor_decl list

| Pdec_open

| Pdec_extend of Path.t Location.loc * ctor_decl list	(*
Internal; this should never be present in the AST.
*)

type pattern = {

pat_desc : pattern_desc;

pat_loc : Location.t;

}

and pattern_desc =

| Ppat_any

| Ppat_variable of Ast_types.str

| Ppat_constraint of pattern * type_expr

| Ppat_tuple of pattern list

| Ppat_or of pattern * pattern

| Ppat_integer of string * char option	(*
Delayed parsing of integers so that we can complain about the right bounds mismatch if they overflow.
*)

| Ppat_literal of Ast_types.literal

| Ppat_record of (Ast_types.lid * pattern) list

| Ppat_ctor of Ast_types.lid * pattern option

| Ppat_row_ctor of Ast_types.str * pattern list

type expression = {

exp_desc : expression_desc;

exp_loc : Location.t;

}

and expression_desc =

| Pexp_apply of expression * (Asttypes.arg_label * expression) list

| Pexp_variable of Ast_types.lid

| Pexp_integer of string * char option	(*
Delayed parsing of integers so that we can complain about the right bounds mismatch if they overflow.
*)

| Pexp_literal of Ast_types.literal

| Pexp_fun of Asttypes.arg_label * pattern * expression * Ast_types.explicitness

| Pexp_newtype of Ast_types.str * expression

| Pexp_seq of expression * expression

| Pexp_let of pattern * expression * expression

| Pexp_instance of Ast_types.str * expression * expression

| Pexp_constraint of expression * type_expr

| Pexp_tuple of expression list

| Pexp_match of expression * (pattern * expression) list

| Pexp_field of expression * Ast_types.lid

| Pexp_record of (Ast_types.lid * expression) list * expression option

| Pexp_ctor of Ast_types.lid * expression option

| Pexp_row_ctor of Ast_types.str * expression list

| Pexp_unifiable of {

mutable expression : expression option;

name : Ast_types.str;

id : int;

}

| Pexp_if of expression * expression * expression option

| Pexp_prover of expression

type conv_type =

| Ptconv_with of Ast_types.mode * type_decl

| Ptconv_to of type_expr

type signature_item = {

sig_desc : signature_desc;

sig_loc : Location.t;

}

and signature = signature_item list

and signature_desc =

| Psig_value of Ast_types.str * type_expr

| Psig_instance of Ast_types.str * type_expr

| Psig_type of type_decl

| Psig_convtype of type_decl * conv_type * Ast_types.str option

| Psig_rectype of type_decl list

| Psig_module of Ast_types.str * module_sig

| Psig_modtype of Ast_types.str * module_sig

| Psig_open of Ast_types.lid

| Psig_typeext of variant * ctor_decl list

| Psig_request of type_expr * ctor_decl

| Psig_multiple of signature

| Psig_prover of signature

| Psig_convert of Ast_types.str * type_expr

and module_sig = {

msig_desc : module_sig_desc;

msig_loc : Location.t;

}

and module_sig_desc =

| Pmty_sig of signature

| Pmty_name of Ast_types.lid

| Pmty_alias of Ast_types.lid

| Pmty_abstract

| Pmty_functor of Ast_types.str * module_sig * module_sig

type statement = {

stmt_desc : statement_desc;

stmt_loc : Location.t;

}

and statements = statement list

and statement_desc =

| Pstmt_value of pattern * expression

| Pstmt_instance of Ast_types.str * expression

| Pstmt_type of type_decl

| Pstmt_convtype of type_decl * conv_type * Ast_types.str option

| Pstmt_rectype of type_decl list

| Pstmt_module of Ast_types.str * module_expr

| Pstmt_modtype of Ast_types.str * module_sig

| Pstmt_open of Ast_types.lid

| Pstmt_open_instance of Ast_types.lid

| Pstmt_typeext of variant * ctor_decl list

| Pstmt_request of type_expr * ctor_decl * (pattern option * expression) option

| Pstmt_multiple of statements

| Pstmt_prover of statements

| Pstmt_convert of Ast_types.str * type_expr

and module_expr = {

mod_desc : module_desc;

mod_loc : Location.t;

}

and module_desc =

| Pmod_struct of statements

| Pmod_name of Ast_types.lid

| Pmod_functor of Ast_types.str * module_sig * module_expr