Module Meja_lib.Path

type t =

| Pident of Ident.t

| Pdot of t * Ast_types.mode * string

| Pocamldot of t * Ast_types.mode * string * string Core_kernel.ref

| Papply of t * t

Paths formed from unique identifiers.

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val pp : Stdlib.Format.formatter -> t -> unit

Pretty print. Identifiers that do not begin with a letter or underscore will be surrounded by parentheses.

val debug_print : Stdlib.Format.formatter -> t -> unit

Debug print. Prints the path with its internal modes and IDs.

val dot : t -> Ident.t -> t

val deconstruct : t -> Ident.t * (Ast_types.mode * string * string Core_kernel.ref option) list

Deconstruct a path to an Ident.t and a list of (mode, name, ocaml name) triples, from innermost to outermost.

For example, deconstruct(A.B.C.d) = (A, [(_, "B", _); (_, "C", _); (_, "d", _)]).

val pdot : t -> t -> t

val of_idents : Ident.t list -> t

Create a path from a list of Ident.ts.

val add_outer_module : Ident.t -> t -> t

Create a new path by prefixing the path with name.

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : (t, comparator_witness) Base__.Comparator.comparator

val validate_lbound : min:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_ubound : max:t Base__.Maybe_bound.t -> t Base__.Validate.check

val validate_bound : min:t Base__.Maybe_bound.t -> max:t Base__.Maybe_bound.t -> t Base__.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : sig ... end

module Set : sig ... end

val to_longident : t -> Ast_types.Longident.t

val to_ocaml_longident : t -> Ast_types.Longident.t

val mode : t -> Ast_types.mode