Module Meja_lib.Patmatch

type pattern_case =

| Pcase_empty

| Pcase_open

| Pcase_type of Type0.type_expr

| Pcase_tuple of pattern_case list

| Pcase_ctor of Ident.t * pattern_case option

| Pcase_literal of Ast_types.literal

| Pcase_record of (pattern_case * Type0.type_expr) Core_kernel.String.Map.t

| Pcase_or of pattern_case list

val pprint_case : Stdlib.Format.formatter -> pattern_case -> Core_kernel__.Import.unit

exception Empty_case

val is_case_empty : pattern_case -> bool

val unifies : (Meja_lib__Envi.t -> Type0.type_expr -> Type0.type_expr -> bool) Core_kernel.ref

val case_of_type_decl : Meja_lib__Envi.t -> Type0.type_expr -> Type0.type_decl -> pattern_case

val case_of_type : Meja_lib__Envi.t -> Type0.type_expr -> pattern_case

val expand_case_of_type : Meja_lib__Envi.t -> Type0.type_expr -> pattern_case

val intersect_case : 'a -> pattern_case -> pattern_case -> pattern_case

val subtract_case : Meja_lib__Envi.t -> pattern_case -> pattern_case -> pattern_case

val case_of_pattern : Envi.t -> Typedast.pattern -> pattern_case

val check_exhaustivity : Envi.t -> Type0.type_expr -> Typedast.pattern Core_kernel.List.t -> pattern_case option

Returns Some pcase for the pattern case that is not matched, if one exists, or None otherwise.

val get_unmatched_cases : count:int -> Envi.t -> Type0.type_expr -> Typedast.pattern Core_kernel.List.t -> pattern_case list