Module Meja_lib.Type0_map

type mapper = {

type_expr : mapper -> Type0.type_expr -> Type0.type_expr;

type_desc : mapper -> Type0.type_desc -> Type0.type_desc;

variant : mapper -> Type0.variant -> Type0.variant;

row : mapper -> Type0.row -> Type0.row;

field_decl : mapper -> Type0.field_decl -> Type0.field_decl;

ctor_args : mapper -> Type0.ctor_args -> Type0.ctor_args;

ctor_decl : mapper -> Type0.ctor_decl -> Type0.ctor_decl;

type_decl : mapper -> Type0.type_decl -> Type0.type_decl;

type_decl_desc : mapper -> Type0.type_decl_desc -> Type0.type_decl_desc;

ident : mapper -> Ident.t -> Ident.t;

path : mapper -> Path.t -> Path.t;

}

val map_list : same:bool Core_kernel.ref -> f:('a -> 'a) -> 'a Core_kernel.List.t -> 'a Core_kernel.List.t

The default mapper (and the functions that it contains) are careful to preserve equality unless some subvalue changes, to ensure that we don't perform unnecessary allocations/GCs during mapping. This also makes it much less likely that we might replace a variant which we planned to mutate later.

CAUTION: In order to break recursion, the default type_expr mapper replaces the type_desc with Treplace typ, where typ is its return value. Any partial-override of this function should ensure that this behaviour is kept using Type1.set_replacement.

To restore the previous values, a snapshot should be taken before with Type1.Snapshot.create, and then a call to Type1.backtrack_replace with the snapshot will undo the Treplace changes.

DO NOT recurse into Treplace values: they will often be self-recursive.

val type_expr : mapper -> Type0.type_expr -> Type0.type_expr

val type_desc : mapper -> Type0.type_desc -> Type0.type_desc

val variant : mapper -> Type0.variant -> Type0.variant

val row : mapper -> Type0.row -> Type0.row

val field_decl : mapper -> Type0.field_decl -> Type0.field_decl

val ctor_args : mapper -> Type0.ctor_args -> Type0.ctor_args

val ctor_decl : mapper -> Type0.ctor_decl -> Type0.ctor_decl

val type_decl : mapper -> Type0.type_decl -> Type0.type_decl

val type_decl_desc : mapper -> Type0.type_decl_desc -> Type0.type_decl_desc

val path : mapper -> Path.t -> Path.t

val ident : mapper -> Ident.t -> Ident.t

val default_mapper : mapper