Module Meja_lib.Type1

type error =

| Mk_wrong_mode of string * Type0.type_expr list * Ast_types.mode * Type0.type_expr

| Mk_invalid of string * Type0.type_expr list * Type0.type_expr

exception Error of Location.t * error

val check_mode : pos:(string * int * int * int) -> error_info:(unit -> string * Type0.type_expr list) -> Ast_types.mode -> Type0.type_expr -> unit

val get_mode : Ast_types.mode -> Type0.type_expr -> Type0.type_expr

Get the stitched type_expr in the given mode.

val type_id : int Core_kernel.ref

val row_id : int Core_kernel.ref

val mk_rp : Type0.row_presence_desc -> Type0.row_presence

val repr : Type0.type_expr -> Type0.type_expr

The representative of a type. This unfolds any Tref values that are present to get to the true underlying type.

val rp_repr : Type0.row_presence -> Type0.row_presence

val rp_strip_subtract : Type0.row_presence -> Type0.row_presence

val set_rp_desc_fwd : (Type0.row_presence -> Type0.row_presence_desc -> unit) Core_kernel.ref

val row_repr_aux : (Meja_lib.Ident.Map.Key.t, Path.t * Type0.row_presence * Type0.type_expr list, Meja_lib.Ident.Map.Key.comparator_witness) Core_kernel.Map.t -> Type0.row -> (Meja_lib.Ident.Map.Key.t, Path.t * Type0.row_presence * Type0.type_expr list, Meja_lib.Ident.Map.Key.comparator_witness) Core_kernel.Map.t * Type0.type_expr * Ast_types.closed_flag

val row_repr_typ_aux : ((Meja_lib.Ident.Map.Key.t, Path.t * Type0.row_presence * Type0.type_expr list, Meja_lib.Ident.Map.Key.comparator_witness) Core_kernel.Map.t * Ast_types.closed_flag) -> Type0.type_expr -> (Meja_lib.Ident.Map.Key.t, Path.t * Type0.row_presence * Type0.type_expr list, Meja_lib.Ident.Map.Key.comparator_witness) Core_kernel.Map.t * Type0.type_expr * Ast_types.closed_flag

val row_repr : Type0.row -> (Meja_lib.Ident.Map.Key.t, Path.t * Type0.row_presence * Type0.type_expr list, Meja_lib.Ident.Map.Key.comparator_witness) Core_kernel.Map.t * Type0.type_expr * Ast_types.closed_flag

val row_repr_typ : Type0.type_expr -> (Meja_lib.Ident.Map.Key.t, Path.t * Type0.row_presence * Type0.type_expr list, Meja_lib.Ident.Map.Key.comparator_witness) Core_kernel.Map.t * Type0.type_expr * Ast_types.closed_flag

val checked_none : Type0.type_expr

An invalid type_expr in checked mode.

val prover_none : Type0.type_expr

An invalid type_expr in prover mode.

val none : Ast_types.mode -> Type0.type_expr

Chose an invalid type_expr based on the given mode.

val other_none : Ast_types.mode -> Type0.type_expr

Chose an invalid type_expr for the other mode to the one given.

other_none mode = none (other_mode mode)

val type_alternate : Type0.type_expr -> Type0.type_expr

val row_alternate : Type0.row -> Type0.row

val is_poly : Type0.type_expr -> bool

val is_valid : Type0.type_expr -> bool

Returns true if the type_expr argument is valid, false otherwise. Can be used to check that a type-stitching has been created or modified correctly.

val is_invalid : Type0.type_expr -> bool

Equivalent to not (is_valid typ).

val equal : Type0.type_expr -> Type0.type_expr -> bool

Judge equality based on type id.

val check_valid : pos:(string * int * int * int) -> error_info:(unit -> string * Type0.type_expr list) -> Type0.type_expr -> unit

val failure_type_id : Core_kernel.Int.t option

val mk' : mode:Ast_types.mode -> int -> Type0.type_desc -> Type0.type_expr

Make a new type at the given mode and depth.

The type_alternate field is set to the invalid value none (other_mode mode).

val stitch : Type0.type_expr -> Type0.type_expr -> Type0.type_expr

stitch typ typ' sets typ and typ' as eachothers type_alternates. Returns the first argument typ.

Raises AssertionError if the types have the same modes, or if either type has an already-initialised type_alternate.

val tri_stitch : Type0.type_expr -> Type0.type_expr -> Type0.type_expr -> Type0.type_expr

tri_stitch in_typ ptyp ctyp stitches the prover-mode type ptyp to the checked-mode ctyp, and sets the type_alternate of the checked-mode type in_typ to ptyp.

This allows in_typ to be the tdec_ret for a type defined in checked mode, so that a round-trip Checked -> Prover -> Checked conversion chain may start and end with different types.

The canonical example of this is bool -> bool -> boolean.

val are_stitched : Type0.type_expr -> Type0.type_expr -> bool

Returns true if the types are stitched together, false otherwise.

val mkvar : mode:Ast_types.mode -> int -> string option -> Type0.type_expr

Create a new type variable with the given name.

The returned type is properly stitched to an equivalent type in the other mode.

module Mk : sig ... end

Constructors for stitched types.

type change =

| Depth of Type0.type_expr * int

| Desc of Type0.type_expr * Type0.type_desc

| Replace of Type0.type_expr * Type0.type_desc

| Row_presence of Type0.row_presence * Type0.row_presence_desc

| Row_replace of Type0.row_presence * Type0.row_presence_desc

val debug_print_change : Stdlib.Format.formatter -> change -> unit

module Snapshot : sig ... end

Implements a weak, mutable linked-list containing the history of changes.

val revert : change -> unit

val backtrack : Snapshot.t -> unit

val filtered_backtrack : f:(change -> bool) -> Snapshot.t -> unit

val fold : init:'a -> f:('a -> Type0.type_expr -> 'a) -> Type0.type_expr -> 'a

val iter : f:(Type0.type_expr -> unit) -> Type0.type_expr -> unit

val copy_desc : f:(Type0.type_expr -> Type0.type_expr) -> Type0.type_desc -> Type0.type_desc

Make a copy of the type_desc, using f as a recursor. Unfolds any references to make a copy of the representative's description.

val equal_at_depth : get_decl:(Path.t -> Type0.type_decl) -> depth:int -> Type0.type_expr -> Type0.type_expr -> bool

val set_depth : int -> Type0.type_expr -> unit

val update_depth : int -> Type0.type_expr -> unit

val unify_depths : Type0.type_expr -> Type0.type_expr -> unit

val set_rp_desc : Type0.row_presence -> Type0.row_presence_desc -> unit

val set_desc : Type0.type_expr -> Type0.type_desc -> unit

val unsafe_set_single_replacement : Type0.type_expr -> Type0.type_expr -> unit

val set_replacement : Type0.type_expr -> Type0.type_expr -> unit

val backtrack_replace : Snapshot.t -> unit

Backtrack only undoing the Replace operations since the last snapshot.

val set_repr : Type0.type_expr -> Type0.type_expr -> unit

set_repr typ typ' sets the representative of typ to be typ'.

val choose_variable_name : Type0.type_expr -> Type0.type_expr -> unit

choose_variable_name var typ lifts the type variable name for var into the type_desc for typ when it is also a variable.

Raises AssertionError if var is not a type variable.

val add_instance : unify:(Type0.type_expr -> Type0.type_expr -> unit) -> Type0.type_expr -> Type0.type_expr -> unit

add_instance var typ' changes the representative of the type variable var to typ'. If typ' is also a type variable, then the user-provided of var is added to typ', unless typ' already has a user-provided name of its own.

Raises AssertionError if var is not a type variable.

val flatten : Type0.type_expr -> Type0.type_expr

Create an equivalent type by unfolding all of the type representatives.

val type_vars : ?depth:int -> Type0.type_expr -> (Type0.type_expr, Typeset.Elt.comparator_witness) Core_kernel.Set.t

val mk_option : (Type0.type_expr -> Type0.type_expr) Core_kernel.ref

val bubble_label_aux : Ast_types.arg_label -> Type0.type_expr -> (Type0.type_expr * Ast_types.explicitness * Ast_types.arg_label) option * Type0.type_expr

val bubble_label : Ast_types.arg_label -> Type0.type_expr -> Type0.type_expr

val discard_optional_labels : Type0.type_expr -> Type0.type_expr

val is_arrow : Type0.type_expr -> bool

val is_var : Type0.type_expr -> bool

val is_replace : Type0.type_expr -> bool

val get_replace : Type0.type_expr -> Type0.type_expr option

val get_rev_arrow_args : Type0.type_expr -> (Type0.type_expr * Ast_types.explicitness * Ast_types.arg_label) list * Type0.type_expr

val get_implicits : Type0.type_expr -> (Ast_types.arg_label * Type0.type_expr) Core_kernel.List.t * Type0.type_expr

val get_rev_implicits : Type0.type_expr -> (Ast_types.arg_label * Type0.type_expr) list * Type0.type_expr

val contains : Type0.type_expr -> in_:Type0.type_expr -> bool

Returns true if typ is a strict subtype of in_ (i.e. excluding typ == in_), or false otherwise.

val get_same_mode : Type0.type_expr -> Type0.type_expr -> Type0.type_expr * Type0.type_expr

Collapse any Tother_mode types until the modes are the same, if possible.

Checked mode is always preferred over prover mode, so that tri-stitching is retained whenever possible.

val remove_opaques : Type0.type_expr -> Type0.type_expr

Remove any Topaque wrappers, returning the outermost non-opaque type.

val remove_mode_changes : Type0.type_expr -> Type0.type_expr

Remove any Tother_mode wrappers.

module Decl : sig ... end

val report_error : Stdlib.Format.formatter -> error -> unit