Module Type1.Snapshot

Implements a weak, mutable linked-list containing the history of changes.

Every change is added to the same list, and the snapshots correspond to cuts of this list. We gain several advantages from using a weak, mutable linked-list: * if there are no active snapshots, the whole list will be GC'd and any changes don't need to be stored at all * if there are active snapshots, OCaml will GC the history of changes up to the first active snapshot * all simultaneously active snapshots point to a part of the same physical list * the snapshots can be erased during backtracking, so that different snapshots can't be used out of order to 'restore' a state that didn't exist previously

type t

val create : unit -> t

Get a new snapshot.

val add_to_history : change -> unit

Add a change to the history of all active snapshots.

val backtrack : t -> change list

Erase the history back to the snapshot, and return the list of changes that occurred since, ordered from newest to oldest.

val filtered_backtrack : f:(change -> bool) -> t -> change list

Erase all changes matching the filter f back to the snapshot, and return the erased changes, ordered from newest to oldest.

val debug_print : Stdlib.Format.formatter -> t -> unit

val debug_print_latest : Stdlib.Format.formatter -> unit -> unit