Module Typet.TypeDecl

val generalise : Parsetypes.type_decl -> Parsetypes.type_decl * Parsetypes.type_decl

val predeclare : Envi.t -> Parsetypes.type_decl -> Envi.t

Add the type to the environment as an abstract type.

This can be used to set up the environment for recursive or co-recursive types, which need to find their own identifier in scope.

val import : ?name:Ident.t -> ?other_name:Path.t -> ?tri_stitched:(Envi.t -> Type0.type_expr list -> Type0.type_expr) -> ?newtype:bool -> Parsetypes.type_decl -> Envi.t -> Typedast.type_decl * Envi.t

Import a type declaration.

The name parameter can be used to specify an existing identifier for the type, instead of generating a fresh one.

If other_name is specified, then the type declaration will be stitched to the type with that name; otherwise, the type is stitched to a type of its own name in the other mode.

Alternatively, if tri_stitched is specified, then it will be evaluated with the environment containing the type parameters of the declaration and the type parameters of the type declaration. The declaration's type will be tri-stitched to its return value as tri-stitched as name -> tri_stitched@{P} <-> tri_stitched@{C}. This argument may only be used in checked mode, and cannot be used in combination with other_name.

val import_convertible : Parsetypes.type_decl -> Parsetypes.conv_type -> Envi.t -> Typedast.type_decl * Typedast.conv_type * Envi.t

Import a type declaration, stitching it to the type described by the conv_type argument.

val import_rec : Parsetypes.type_decl list -> Envi.t -> Typedast.type_decl list * Envi.t

Import the given type declarations recursively.